+++ packages/ssz-type-schema/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.1",
   "description": "Utility methods and types for describing an ssz object",
   "author": "ChainSafe Systems",
-  "license": "LGPL-3.0",
+  "license": "Apache-2.0",
   "bugs": {
     "url": "https://github.com/ChainSafe/lodestar/issues"
   },
